แก้ components 80%
