+++ src/components/Admin/Button/UserAction.tsx
@@ -1,4 +1,4 @@
-import { BlockIcon, UnblockIcon, PersonIcon, MoreIcon, BinIcon, AcceptIcon, RejectIcon } from "../../../../shared/Icon";
+import { BlockIcon, UnblockIcon, PersonIcon, MoreIcon, BinIcon, AcceptIcon, RejectIcon } from "@/shared/Icon";
 
 export function DefaultAction() {
     return (
