+++ src/components/Admin/Card/UserCard.tsx
@@ -9,7 +9,7 @@ export default function UserCard({ src, firstname, lastname }: {
         <div className="flex w-[270px] h-[72px] pl-[16px] pr-[16px]">
             <div className="flex justify-between gap-[16px] items-center">
                 {src && <Image src={src} alt="Failed" width={40} height={40} className="rounded-[20px] w-[40px] h-[40px] object-cover" />}
-                <div className="text-standard-gray w-[159px] leading-[1]">{firstname} {lastname}</div>
+                <div className="text-standard-gray text-body w-[159px] leading-[1]">{firstname} {lastname}</div>
             </div>
         </div>
     )
