+++ src/shared/Icon.tsx
@@ -280,3 +280,163 @@ export const BillIcon = () => {
     </svg>
   );
 };
+
+export function NullCheckboxIcon() {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M1.77778 0.5H14.2222C14.5611 0.5 14.8861 0.634623 15.1257 0.874252C15.3654 1.11388 15.5 1.43889 15.5 1.77778V14.2222C15.5 14.5611 15.3654 14.8861 15.1257 15.1257C14.8861 15.3654 14.5611 15.5 14.2222 15.5H1.77778C1.43889 15.5 1.11388 15.3654 0.874252 15.1257C0.634623 14.8861 0.5 14.5611 0.5 14.2222V1.77778C0.5 1.06725 1.06725 0.5 1.77778 0.5Z" stroke="#B3B3B3" />
+    </svg>
+
+  );
+}
+
+export function DropdownArrowIcon() {
+  return (
+    <svg
+      width="14"
+      height="7"
+      viewBox="0 0 14 7"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M12.2807 0.966797L7.93404 5.31346C7.4207 5.8268 6.5807 5.8268 6.06737 5.31346L1.7207 0.966797" stroke="#333333" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round" />
+    </svg>
+  );
+}
+
+export function BlockIcon() {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M8 0C3.584 0 0 3.584 0 8C0 12.416 3.584 16 8 16C12.416 16 16 12.416 16 8C16 3.584 12.416 0 8 0ZM1.6 8C1.6 4.464 4.464 1.6 8 1.6C9.48 1.6 10.84 2.104 11.92 2.952L2.952 11.92C2.0733 10.8026 1.59699 9.42153 1.6 8ZM8 14.4C6.52 14.4 5.16 13.896 4.08 13.048L13.048 4.08C13.9267 5.19743 14.403 6.57847 14.4 8C14.4 11.536 11.536 14.4 8 14.4Z" fill="#E64D4D" />
+    </svg>
+  );
+}
+
+export function UnblockIcon() {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M11.0002 5.59945H4.59982V3.99937C4.59865 3.52429 4.7385 3.05956 5.00165 2.66402C5.2648 2.26848 5.63941 1.95993 6.07804 1.77746C6.51668 1.59498 6.9996 1.54678 7.46565 1.63896C7.93169 1.73114 8.3599 1.95955 8.69604 2.29528C8.99622 2.60299 9.21126 2.98345 9.32007 3.39933C9.34634 3.50125 9.39242 3.59699 9.45568 3.68109C9.51895 3.76519 9.59816 3.83601 9.6888 3.8895C9.77943 3.94299 9.87971 3.9781 9.98392 3.99283C10.0881 4.00757 10.1942 4.00163 10.2961 3.97536C10.398 3.9491 10.4938 3.90302 10.5779 3.83975C10.662 3.77649 10.7328 3.69727 10.7863 3.60664C10.8398 3.516 10.8749 3.41572 10.8896 3.31152C10.9044 3.20731 10.8984 3.10122 10.8722 2.99931C10.6895 2.30726 10.3281 1.67542 9.8241 1.16722C9.26421 0.609079 8.5515 0.229356 7.77596 0.0759931C7.00041 -0.0773693 6.19681 0.00250687 5.46663 0.305536C4.73645 0.608564 4.11244 1.12116 3.67338 1.77859C3.23432 2.43602 2.9999 3.2088 2.99974 3.99937V5.59945C2.36318 5.59945 1.7527 5.85232 1.30259 6.30243C0.852479 6.75254 0.599609 7.36302 0.599609 7.99958V13.5999C0.599609 14.2364 0.852479 14.8469 1.30259 15.297C1.7527 15.7471 2.36318 16 2.99974 16H11.0002C11.6367 16 12.2472 15.7471 12.6973 15.297C13.1474 14.8469 13.4003 14.2364 13.4003 13.5999V7.99958C13.4003 7.36302 13.1474 6.75254 12.6973 6.30243C12.2472 5.85232 11.6367 5.59945 11.0002 5.59945ZM11.8002 13.5999C11.8002 13.8121 11.7159 14.0156 11.5659 14.1656C11.4158 14.3156 11.2123 14.3999 11.0002 14.3999H2.99974C2.78755 14.3999 2.58406 14.3156 2.43402 14.1656C2.28398 14.0156 2.19969 13.8121 2.19969 13.5999V7.99958C2.19969 7.78739 2.28398 7.5839 2.43402 7.43386C2.58406 7.28382 2.78755 7.19953 2.99974 7.19953H11.0002C11.2123 7.19953 11.4158 7.28382 11.5659 7.43386C11.7159 7.5839 11.8002 7.78739 11.8002 7.99958V13.5999Z" fill="#333333" />
+    </svg>
+  );
+}
+
+export function PersonIcon() {
+  return (
+    <svg
+      width="18"
+      height="18"
+      viewBox="0 0 18 18"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M1 15C1 13.9391 1.42143 12.9217 2.17157 12.1716C2.92172 11.4214 3.93913 11 5 11H13C14.0609 11 15.0783 11.4214 15.8284 12.1716C16.5786 12.9217 17 13.9391 17 15C17 15.5304 16.7893 16.0391 16.4142 16.4142C16.0391 16.7893 15.5304 17 15 17H3C2.46957 17 1.96086 16.7893 1.58579 16.4142C1.21071 16.0391 1 15.5304 1 15Z" stroke="#333333" stroke-width="1.5" stroke-linejoin="round" />
+      <path d="M9 7C10.6569 7 12 5.65685 12 4C12 2.34315 10.6569 1 9 1C7.34315 1 6 2.34315 6 4C6 5.65685 7.34315 7 9 7Z" stroke="#333333" stroke-width="1.5" />
+    </svg>
+  );
+}
+
+export function MoreIcon() {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M10.5 8.25V1.5H3.75V0H12V8.25H10.5ZM6.75 12V5.25H0V3.75H8.25V12H6.75Z" fill="#333333" />
+    </svg>
+  );
+}
+
+export function BinIcon() {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path fill-rule="evenodd" clip-rule="evenodd" d="M4.75 3V1.5H9.25V3H4.75ZM3.25 3V1C3.25 0.734784 3.35536 0.48043 3.54289 0.292893C3.73043 0.105357 3.98478 0 4.25 0L9.75 0C10.0152 0 10.2696 0.105357 10.4571 0.292893C10.6446 0.48043 10.75 0.734784 10.75 1V3H13.25C13.4489 3 13.6397 3.07902 13.7803 3.21967C13.921 3.36032 14 3.55109 14 3.75C14 3.94891 13.921 4.13968 13.7803 4.28033C13.6397 4.42098 13.4489 4.5 13.25 4.5H12.885L12.142 14.153C12.1034 14.6554 11.8767 15.1248 11.507 15.4673C11.1373 15.8097 10.6519 16 10.148 16H3.852C3.34808 16 2.86274 15.8097 2.49304 15.4673C2.12335 15.1248 1.89655 14.6554 1.858 14.153L1.115 4.5H0.75C0.551088 4.5 0.360322 4.42098 0.21967 4.28033C0.0790177 4.13968 0 3.94891 0 3.75C0 3.55109 0.0790177 3.36032 0.21967 3.21967C0.360322 3.07902 0.551088 3 0.75 3H3.25ZM2.62 4.5H11.38L10.646 14.038C10.6364 14.1636 10.5798 14.2809 10.4875 14.3666C10.3952 14.4522 10.2739 14.4999 10.148 14.5H3.852C3.72607 14.4999 3.60482 14.4522 3.5125 14.3666C3.42017 14.2809 3.36357 14.1636 3.354 14.038L2.62 4.5Z" fill="#E64D4D" />
+    </svg>
+  );
+}
+
+export function AcceptIcon() {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M4.93925 9.70027L1.62542 6.38644C1.44686 6.20788 1.20467 6.10756 0.952148 6.10756C0.699623 6.10756 0.45744 6.20788 0.278878 6.38644C0.100315 6.565 0 6.80719 0 7.05971C0 7.18475 0.0246282 7.30856 0.0724782 7.42408C0.120328 7.5396 0.190463 7.64457 0.278878 7.73298L4.27075 11.7249C4.6432 12.0973 5.24484 12.0973 5.61729 11.7249L15.7211 1.62102C15.8997 1.44246 16 1.20028 16 0.947754C16 0.695228 15.8997 0.453046 15.7211 0.274483C15.5426 0.095921 15.3004 -0.00439453 15.0479 -0.00439453C14.7953 -0.00439453 14.5531 0.095921 14.3746 0.274483L4.93925 9.70027Z" fill="#05CE78" />
+    </svg>
+  );
+}
+
+export function RejectIcon() {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M13 1L1 13M1 1L13 13" stroke="#E64D4D" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" />
+    </svg>
+  );
+}
+
+export function OptionIcon() {
+  return (
+    <svg
+      width="18"
+      height="4"
+      viewBox="0 0 18 4"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M1.5 0.5C0.675 0.5 0 1.175 0 2C0 2.825 0.675 3.5 1.5 3.5C2.325 3.5 3 2.825 3 2C3 1.175 2.325 0.5 1.5 0.5ZM16.5 0.5C15.675 0.5 15 1.175 15 2C15 2.825 15.675 3.5 16.5 3.5C17.325 3.5 18 2.825 18 2C18 1.175 17.325 0.5 16.5 0.5ZM9 0.5C8.175 0.5 7.5 1.175 7.5 2C7.5 2.825 8.175 3.5 9 3.5C9.825 3.5 10.5 2.825 10.5 2C10.5 1.175 9.825 0.5 9 0.5Z" fill="#808080" />
+    </svg>
+
+  );
+}
+
+export function FilterIcon() {
+  return (
+    <svg
+      width="26"
+      height="16"
+      viewBox="0 0 26 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M0 1.21212C0 0.54303 0.54303 0 1.21212 0H24C24.3215 0 24.6298 0.127705 24.8571 0.355022C25.0844 0.582339 25.2121 0.890647 25.2121 1.21212C25.2121 1.5336 25.0844 1.8419 24.8571 2.06922C24.6298 2.29654 24.3215 2.42424 24 2.42424H1.21212C0.54303 2.42424 0 1.88121 0 1.21212ZM3.87879 8C3.87879 7.33091 4.42182 6.78788 5.09091 6.78788H20.1212C20.4427 6.78788 20.751 6.91558 20.9783 7.1429C21.2056 7.37022 21.3333 7.67853 21.3333 8C21.3333 8.32148 21.2056 8.62978 20.9783 8.8571C20.751 9.08442 20.4427 9.21212 20.1212 9.21212H5.09091C4.42182 9.21212 3.87879 8.66909 3.87879 8ZM8.9697 13.5758C8.64822 13.5758 8.33992 13.7035 8.1126 13.9308C7.88528 14.1581 7.75758 14.4664 7.75758 14.7879C7.75758 15.1094 7.88528 15.4177 8.1126 15.645C8.33992 15.8723 8.64822 16 8.9697 16H16.2424C16.5639 16 16.8722 15.8723 17.0995 15.645C17.3268 15.4177 17.4545 15.1094 17.4545 14.7879C17.4545 14.4664 17.3268 14.1581 17.0995 13.9308C16.8722 13.7035 16.5639 13.5758 16.2424 13.5758H8.9697Z" fill="#B3B3B3" />
+    </svg>
+
+  );
+}
+
+export function SearchIcon() {
+  return (
+    <svg
+      width="18"
+      height="18"
+      viewBox="0 0 18 18"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg">
+      <path d="M13.4444 13.4444L17 17M1 8.11111C1 9.9971 1.7492 11.8058 3.0828 13.1394C4.41639 14.473 6.22513 15.2222 8.11111 15.2222C9.9971 15.2222 11.8058 14.473 13.1394 13.1394C14.473 11.8058 15.2222 9.9971 15.2222 8.11111C15.2222 6.22513 14.473 4.41639 13.1394 3.0828C11.8058 1.7492 9.9971 1 8.11111 1C6.22513 1 4.41639 1.7492 3.0828 3.0828C1.7492 4.41639 1 6.22513 1 8.11111Z" stroke="#1D9BF0" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" />
+    </svg>
+
+  );
+}
